requirements.txt
PyDejavu
pyaudio
mysqldb
pydub
numpy
scipy
